# src/integration-test.disabled/resources/api/openapi.yaml
openapi: 3.1.0
info:
  title: rtj-management
  contact: {}
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: "1.0"
servers:
- url: http://localhost:59560
  description: Generated server url
...
paths:
...
  /{municipalityId}/{namespace}/errands/{errandId}/decisions:
    get:
      tags:
      - Errand decisions
      summary: List decisions
      operationId: readDecisions
      parameters:
      - name: municipalityId
        in: path
        required: true
        schema:
          type: string
        example: 2281
      - name: namespace
        in: path
        required: true
        schema:
          type: string
          pattern: "[\\w|\\-]+"
        example: MY_NAMESPACE
      - name: errandId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successful Operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Decision"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    post:
      tags:
      - Errand decisions
      summary: Create decision
      operationId: createDecision
      parameters:
      - name: municipalityId
        in: path
        required: true
        schema:
          type: string
        example: 2281
      - name: namespace
        in: path
        required: true
        schema:
          type: string
          pattern: "[\\w|\\-]+"
        example: MY_NAMESPACE
      - name: errandId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Decision"
        required: true
      responses:
        "201":
          description: Successful operation
          headers:
            Location:
              style: simple
              schema:
                type: string
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
  /{municipalityId}/{namespace}/errands/{errandId}/decisions/{decisionId}:
    get:
      tags:
      - Errand decisions
      summary: Read decision
      operationId: readDecision
      parameters:
      - name: municipalityId
        in: path
        required: true
        schema:
          type: string
        example: 2281
      - name: namespace
        in: path
        required: true
        schema:
          type: string
          pattern: "[\\w|\\-]+"
        example: MY_NAMESPACE
      - name: errandId
        in: path
        required: true
        schema:
          type: string
      - name: decisionId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Decision"
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    delete:
      tags:
      - Errand decisions
      summary: Delete decision
      operationId: deleteDecision
      parameters:
      - name: municipalityId
        in: path
        required: true
        schema:
          type: string
        example: 2281
      - name: namespace
        in: path
        required: true
        schema:
          type: string
          pattern: "[\\w|\\-]+"
        example: MY_NAMESPACE
      - name: errandId
        in: path
        required: true
        schema:
          type: string
      - name: decisionId
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: Successful operation
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
components:
  schemas:
    Problem:
      type: object
      properties:
        detail:
          type: string
        title:
          type: string
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        status:
          type: integer
          format: int32
    ConstraintViolationProblem:
      type: object
      properties:
        type:
          type: string
          format: uri
        status:
          type: integer
          format: int32
        violations:
          type: array
          items:
            $ref: "#/components/schemas/Violation"
        title:
          type: string
        detail:
          type: string
        causeAsProblem:
          $ref: "#/components/schemas/ThrowableProblem"
        instance:
          type: string
          format: uri
    ThrowableProblem:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        causeAsProblem: {}
    Violation:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
...
    Decision:
      type: object
      description: "Decision recorded against an errand. Both system-generated decisions\
        \ (e.g. a DMN-evaluated recommendation produced by a BPMN process) and human\
        \ decisions (e.g. a handläggare approving a payment) are stored here, distinguished\
        \ by `decisionType`. The list on the errand grows over time and is the audit\
        \ trail of every decision made on the case."
      properties:
        id:
          type: string
          description: Unique identifier
          examples:
          - cb20c51f-fcf3-42c0-b613-de563634a8ec
          readOnly: true
        decisionType:
          type: string
          description: "Decision category. Free-form string; conventionally `RECOMMENDATION`\
            \ for DMN-produced suggestions and `PAYMENT` for handläggare APPROVE/REJECT\
            \ decisions, but namespaces are encouraged to define their own."
          examples:
          - PAYMENT
        value:
          type: string
          description: Decision value. For binary outcomes use `APPROVED`/`REJECTED`;
            for richer outputs (e.g. a calculated amount) use the value itself or
            a short label.
          examples:
          - APPROVED
        description:
          type: string
          description: Optional human-readable description or motivation for the decision
          examples:
          - "Beslutsförslag enligt regelverk: 7900 kr, ingen varning"
        createdBy:
          type: string
          description: "Identifier of the actor that produced the decision. Use the\
            \ handläggare userId for human decisions or a system identifier (e.g.\
            \ `operaton`, `dmn-engine`) for automated ones."
          examples:
          - jane01doe
        created:
          type: string
          format: date-time
          description: Timestamp the decision was recorded (server-assigned)
          readOnly: true
...
  securitySchemes: {}